+++ main.cpp
@@ -4,7 +4,7 @@ using namespace std;
 
 #include "Analyzer.h"
 
-bool inRange(const Vec2& vec, float mn, float mx) {
+bool inRange(const Vec2& vec, double mn, double mx) {
     return vec.x >= mn && vec.x <= mx && vec.y >= mn && vec.y <= mx;
 }
 
@@ -19,12 +19,12 @@ void runTests(const Vec2& mapDims, const vector<unsigned char>& beforeData, cons
         cout << "P1 " << p1.x << "," << p1.y << endl;
         cout << "P2 " << p2.x << "," << p2.y << endl;
 
-        float d01 = Analyzer::CalculatePathLength(beforeData, mapDims, p0,p1);
-        float d12 = Analyzer::CalculatePathLength(beforeData, mapDims, p1,p2);
-        float d02 = Analyzer::CalculatePathLength(beforeData,mapDims,p2,p0);
+        double d01 = Analyzer::CalculatePathLength(beforeData, mapDims, p0,p1);
+        double d12 = Analyzer::CalculatePathLength(beforeData, mapDims, p1,p2);
+        double d02 = Analyzer::CalculatePathLength(beforeData,mapDims,p2,p0);
 
         cout << "d01 = " << d01 << " d12 = " << d12 << " d02 = " << d02 << endl;
-        float error = d02 - d01 - d12;
+        double error = d02 - d01 - d12;
         cout << "error = " << error << endl;
     }
 
@@ -33,18 +33,18 @@ void runTests(const Vec2& mapDims, const vector<unsigned char>& beforeData, cons
     {
         Vec2 start(0,0);
         Vec2 end(511,0);
-        float cp0 = 0;
-        float cp1 = 0;
+        double cp0 = 0;
+        double cp1 = 0;
         for (int i=start.x + 1; i <= end.x; ++i) {
-            float d0 = 11.0f * abs(beforeData[i] - beforeData[i-1]);
+            double d0 = 11.0f * abs(beforeData[i] - beforeData[i-1]);
             cp0 += sqrt(30*30+d0*d0);
-            float d1 = 11.0f * abs(afterData[i] - afterData[i-1]);
+            double d1 = 11.0f * abs(afterData[i] - afterData[i-1]);
             cp1 += sqrt(30*30 + d1*d1);
         }
 
         cout << " Horizontal test\n: Before: " << cp0 << " after " << cp1 << " delta " << (cp1 - cp0) << endl;
-        float ap0 = Analyzer::CalculatePathLength(beforeData,mapDims,start, end);
-        float ap1 = Analyzer::CalculatePathLength(afterData,mapDims,start,end);
+        double ap0 = Analyzer::CalculatePathLength(beforeData,mapDims,start, end);
+        double ap1 = Analyzer::CalculatePathLength(afterData,mapDims,start,end);
         cout << "Calculated:\n  Before: " << ap0 << " after " << ap1 << " delta " << ap1 - ap0 << endl;
 
         cout << "Compare Horizontal test from " << start.x << "," << start.y << " to " << end.x << "," << end.y << endl;
@@ -86,8 +86,8 @@ int main(int argc, char** argv)
             continue;
         }
         cout << "Calculating each path separately" << endl;
-        float preDistance = Analyzer::CalculatePathLength(beforeData, mapDims, start,end);
-        float postDistance = Analyzer::CalculatePathLength(afterData, mapDims,start,end);
+        double preDistance = Analyzer::CalculatePathLength(beforeData, mapDims, start,end);
+        double postDistance = Analyzer::CalculatePathLength(afterData, mapDims,start,end);
         cout << "Before eruption: " << preDistance << endl;
         cout << "After eruption: " << postDistance << endl;
         cout << "Delta: " << postDistance - preDistance << endl;
